* Package.pm: Write the new index atomicly to avoid a corrupted 	index should fink be killed in the middle of reindexing.
Ben Hines noticed that killing fink in the middle of a reindexing left him
with a mangled index.  It may have been due to the fact that we're
overwriting the original index non-atomicly.  Not sure if that was the
problem, but it can't hurt to play it safe and write to a temp file then
rename().