Add a TODO note to the Shlibs.m code
There already is a little bit of RuntimeDepends code in there, put in
there by somebody else in the past. However, it seems to be incomplete
and possibly partially incorrect. Alas, the code in Shlibs.pm seems to
be completely unused at the time, so I am not worrying about this for
now.