Use File::Temp::tempfile() instead of tmpnam() for temporary script fles.
Based (with minor alteration) on Fink patch tracker item #3294598 by
dustinac, which provides the following explanation.

"The fink program saves scripts it needs to execute into /var/tmp. At
least if you interrupt builds as much as I do, these temporary files
tend to accumulate in /var/tmp over the years. This patch changes
Services.pm to instead put the scripts in /tmp. The advantage is that
/tmp is not preserved across reboots, so the scripts won't accumulate in
the same way."