Remove xft2* and fontconfig2* from x11* (virt-pkg) Provides lists
Fink has real xft2* and fontconfig2* packages that have been in the
default search paths for a long time, and therefore likely mask the
ones in X11's paths. The versions and install_names are not the same,
so they are not drop-in replacements. It's likely a nondeterministic
build process unless it forces which of them to use and fink's paths
usually take precedence over system's, so fink-package-precedence
rejects the x11 ones anyway. Therefore, it's not sufficient (or
consistent) for a *Depends to be satisfied by the x11 Provides: of
them.