fix .deb test for perl compiled C modules in non-perlversioned location
By fink packaging policy, perl modules containing compiled C (.bundle
files) must not be stored in perl-version-agnostic part of INC because
they are tied to the perl version API/ABI. The test had hardcoded the
perl-arch as the one used prior to perl581 (!), which changed at that
point. The path has a well-defined format, so let's just not hardcode
it at all.