HACK: Fix bootstrap issue by disabling NLS in dpkg.
From J. Howarth.  I'm not sure what the ramifications of disabling NLS ultimately are.

Justin says:
it fails when it tried to extract the deb that are made after the final
dpkg is installed but before bootstrap is done. once bootstrap is done
and it has failed you can just rebuild the failed deb and it works fine
(the only deb it happens in is apt because it's made after dpkg during
bootstrapping) outside of bootstrapping it's fine

Max adds: from what I understand, it is currently not known *why*
disabling NLS fixes this problem, nor what the real reason for that is.
Since NLS is probably important for many of our users, we should really
find out the real cause of this problem and fix its root cause. Then
again, as long as dpkg is the last thing we build during bootstrap, we
seem to be safe. So we could just ignore it for now.