perlmods: generate symbols file based on shlibs, for automatic dependencies
dpkg-shlibdep uses the symbols file to generate auto shlibs aka auto
depends. I worked very hard on this part and it took lots of work to get
this working, but it does. Basically one can put ${Shlibs} in the deb
file and once the work is completed dpkg-shlibdep would replace that
line in the control file with the actual deps based on the symbols used
and the symbols in the symbols files.

For a few examples, see:

https://wiki.debian.org/UsingSymbolsFiles
http://www.debian.org/doc/manuals/maint-guide/advanced.en.html

See also the last paragraph of this section of the Debian Policy Manual:

http://www.debian.org/doc/debian-policy/ch-sharedlibs.html#s-sharedlibs-depends