perlmods: fix .la files during build time, to allow debsums support
The debsums command (part of dpkg) compares the MD5 checksum of all
files installed via .deb files into your Fink tree to the corresponding
checksum derived from the .deb file. This can be used to perform an
automatic integrity check of your system, checking if the files on your
file system have the correct checksum. If not, then something or
somebody must have modified the file since it was installed (e.g. an
intruder on your system).

For this it is important that files are not modified after installation.
But we did exactly that with .la files: We fixed them up in a postinst
script. Since this would break the md5sums mechanism, we now perform
this fixup already while building the .deb file.