bootstrap: make the 64bit dist the default on 10.6
I think I remember this being discussed previously. According to pdb, the 32bit dist for 10.6 has 12188 packages, while the 64bit one has 11792 packages, which is a difference of only 396 packages. Or, in other words, the 64bit dist has about 97% of the packages as the 32bit one on 10.6. Changing the default to 64bit on 10.6 would make portability with newer versions of OS X easier, because then the architecture would be one less thing to worry about.